require root

+++ cleanup-crap.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ `whoami` != 'root' ]
+  then
+    echo "You must be root to do this."
+    exit
+fi
+
 # delete adobe launch agents
 rm /Library/LaunchAgents/com.adobe.*
 rm /Library/LaunchDaemons/com.adobe.*
